\name{pumilioR}

\alias{pumilioR}

\docType{package}

\title{pumilioR}

\description{Functions to get data from a Pumilio sound archive system (http://pumilio.sourceforge.net). 
	Requires version 2.6.0, or newer, of Pumilio installed and XML turned on in the Admin menu.}

\details{
	\tabular{ll}{
		Package: \tab pumilioR\cr
		Type: \tab Package\cr
		Version: \tab 1.0\cr
		Date: \tab 2013-07-07\cr
		License: \tab GPLv3\cr
		}
	}
\author{
Luis J. Villanueva-Rivera
}

\keyword{ pumilioR }
