\name{getVersion}

\alias{getVersion}

\title{Get Pumilio Version}

\description{Function to find the version of the Pumilio system because version 2.6.0, or newer, is required.}

\usage{
	getVersion(pumilio_URL)
	}

\arguments{
	\item{pumilio_URL}{ URL of the Pumilio installation, for example: "http://server.com/pumilio/" }
	}

\value{
	Returns a boolean if the Pumilio version matches the requirement.}

%\author{Luis J. Villanueva-Rivera}

%\examples{
%	}

%\references{
 %}

\keyword{pumilioR}
\keyword{getVersion}
