\name{pumilioR}

\docType{package}

\title{pumilioR}

\description{Functions to get data from a Pumilio system. Requires version 2.6.0, or newer,
of Pumilio installed and XML turned on.}

\details{
	\tabular{ll}{
		Package: \tab pumilioR\cr
		Type: \tab Package\cr
		Version: \tab 0.1\cr
		Date: \tab 2013-06-29\cr
		License: \tab GPLv3\cr
		}
	}
\author{
Luis J. Villanueva-Rivera
}

\keyword{ pumilioR }
